requirements-dev.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
boto3==1.14.0
botocore==1.17.0
celery==4.4.5
Django==3.0.7
django-celery-results==1.2.1
django-redis-cache==2.1.1
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.6.1
jmespath==0.10.0
kombu==4.6.10
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
